[REFACTOR] Renamed package
The package is renamed to reflect the new package structure in interfaces.
